Refactor FileFetch & FetchInterface
- Move FetchInterface & FileFetch sources to frameio & make corresp. source group in CMake
- Refactor FileFetch to VolumeDeformFileFetch (making name more specific to what it does)
