+++ core/SurfelWarpSerial.h
@@ -7,7 +7,7 @@
 #include "common/Constants.h"
 #include "common/ConfigParser.h"
 #include "imgproc/ImageProcessor.h"
-#include "imgproc/FileFetch.h"
+#include "imgproc/frameio/VolumeDeformFileFetch.h"
 #include "core/SurfelGeometry.h"
 #include "core/WarpField.h"
 #include "core/Camera.h"
