+++ imgproc/frameio/VolumeDeformFileFetch.cpp
@@ -1,32 +1,32 @@
 #include <opencv2/opencv.hpp>
 #include "common/OpenCV_CrossPlatform.h"
-#include "imgproc/FileFetch.h"
+#include "VolumeDeformFileFetch.h"
 
-void surfelwarp::FileFetch::FetchDepthImage(size_t frame_idx, cv::Mat & depth_img)
+void surfelwarp::VolumeDeformFileFetch::FetchDepthImage(size_t frame_idx, cv::Mat & depth_img)
 {
 	path file_path = FileNameSurfelWarp(frame_idx, true);
 	//Read the image
 	depth_img = cv::imread(file_path.string(), CV_ANYCOLOR | CV_ANYDEPTH);
 }
 
-void surfelwarp::FileFetch::FetchDepthImage(size_t frame_idx, void * depth_img)
+void surfelwarp::VolumeDeformFileFetch::FetchDepthImage(size_t frame_idx, void * depth_img)
 {
 
 }
 
-void surfelwarp::FileFetch::FetchRGBImage(size_t frame_idx, cv::Mat & rgb_img)
+void surfelwarp::VolumeDeformFileFetch::FetchRGBImage(size_t frame_idx, cv::Mat & rgb_img)
 {
 	path file_path = FileNameSurfelWarp(frame_idx, false);
 	//Read the image
 	rgb_img = cv::imread(file_path.string(), CV_ANYCOLOR | CV_ANYDEPTH);
 }
 
-void surfelwarp::FileFetch::FetchRGBImage(size_t frame_idx, void * rgb_img)
+void surfelwarp::VolumeDeformFileFetch::FetchRGBImage(size_t frame_idx, void * rgb_img)
 {
 
 }
 
-boost::filesystem::path surfelwarp::FileFetch::FileNameVolumeDeform(size_t frame_idx, bool is_depth_img) const
+boost::filesystem::path surfelwarp::VolumeDeformFileFetch::FileNameVolumeDeform(size_t frame_idx, bool is_depth_img) const
 {
 	//Construct the file_name
 	char frame_idx_str[20];
@@ -48,6 +48,6 @@ boost::filesystem::path surfelwarp::FileFetch::FileNameVolumeDeform(size_t frame
 
 
 
-boost::filesystem::path surfelwarp::FileFetch::FileNameSurfelWarp(size_t frame_idx, bool is_depth_img) const {
+boost::filesystem::path surfelwarp::VolumeDeformFileFetch::FileNameSurfelWarp(size_t frame_idx, bool is_depth_img) const {
 	return FileNameVolumeDeform(frame_idx, is_depth_img);
 }
